+++ src/main/java/inu/codin/codin/domain/post/domain/poll/controller/PollController.java
@@ -3,7 +3,7 @@ package inu.codin.codin.domain.post.domain.poll.controller;
 import inu.codin.codin.common.response.SingleResponse;
 import inu.codin.codin.domain.post.domain.poll.dto.PollCreateRequestDTO;
 import inu.codin.codin.domain.post.domain.poll.dto.PollVotingRequestDTO;
-import inu.codin.codin.domain.post.domain.poll.service.PollService;
+import inu.codin.codin.domain.post.domain.poll.service.PollCommandService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -18,14 +18,14 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class PollController {
 
-    private final PollService pollService;
+    private final PollCommandService pollCommandService;
 
     @Operation(summary = "투표 생성")
     @PostMapping
     public ResponseEntity<SingleResponse<?>> createPoll(
             @Valid @RequestBody PollCreateRequestDTO pollRequestDTO) {
 
-        pollService.createPoll(pollRequestDTO);
+        pollCommandService.createPoll(pollRequestDTO);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(new SingleResponse<>(201, "투표 생성 완료", null));
     }
@@ -36,7 +36,7 @@ public class PollController {
             @PathVariable String postId,
             @Valid @RequestBody PollVotingRequestDTO pollRequestDTO) {
 
-        pollService.votingPoll(postId, pollRequestDTO);
+        pollCommandService.votingPoll(postId, pollRequestDTO);
         return ResponseEntity.status(HttpStatus.CREATED)
                 .body(new SingleResponse<>(200, "투표 실시 완료", null));
     }
@@ -44,7 +44,7 @@ public class PollController {
     @Operation(summary = "투표 취소")
     @DeleteMapping("/voting/{postId}")
     public ResponseEntity<SingleResponse<?>> deleteVoting(@PathVariable String postId){
-        pollService.deleteVoting(postId);
+        pollCommandService.deleteVoting(postId);
         return ResponseEntity.ok()
                 .body(new SingleResponse<>(200, "투표 취소 완료", null));
     }
